Extract renderTransformedImage helper for canvas transform rendering
Moved the canvas transform rendering logic into reusable helper functions
in stage-helpers.ts:
- transformSwapsDimensions(): checks if a transform swaps width/height
- renderTransformedImage(): renders an image with transform to a canvas

This simplifies the drag image fix in actor-sprite.tsx and provides
reusable utilities for other canvas rendering that needs transforms.
